requirements.txt
streamlit
yfinance
plotly
pandas
python-dateutil